+++ save_config.py
@@ -8,10 +8,16 @@ os.rename(config_file, config_backup)
 # Write default data to new testing config
 config = open(config_file, 'x')
 config.write("""{
-  "output_dir": "/Users/ilesh/Library/ApplicationSupport/minecraft/mods",
-  "game_version": "1.17.1",
-  "mod_loader": "fabric",
-  "mod_ids": [],
-  "repos": []
+  "active_profile": 0,
+  "profiles": [
+    {
+      "name": "test",
+      "output_dir": "/Users/ilesh/Library/ApplicationSupport/minecraft/mods",
+      "game_version": "1.18.1",
+      "mod_loader": "fabric",
+      "mod_ids": [],
+      "repos": []
+    }
+  ]
 }""")
 config.close()